backend/pyproject.toml
[project]
name = "upmon-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "asyncpg",
    "pydantic-settings",
]

[project.scripts]
upmon-backend = "upmon_backend.main:main"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
